taxi_v4/requirements.txt
gymnasium==1.3.0
matplotlib==3.10.9
numpy==2.4.4
Pillow==12.2.0
